+++ tests/apps/core/test_web_analytics.py
@@ -43,13 +43,13 @@ class WebAnalyticsTestCase(CMSTestCase):
 
         self.assertContains(
             response,
-            "googletagmanager",
-            msg_prefix="Page should include the Google Analytics snippet code",
+            "google-analytics.com",
+            msg_prefix="Page should include the Google Analytics js code",
         )
         self.assertContains(
             response,
             "UA-XXXXXXXXX-X",
-            msg_prefix="Page should include the Google Analytics tracking code",
+            msg_prefix="Page should include the Google Analytics tracking id code",
         )
         self.assertRegex(
             response.content.decode("UTF-8"),
@@ -64,7 +64,7 @@ class WebAnalyticsTestCase(CMSTestCase):
         response_content = response.content.decode("UTF-8")
         self.assertGreater(
             response_content.index("<body"),
-            response_content.index("googletagmanager"),
+            response_content.index("google-analytics.com"),
             "Web tracking should be at the bottom of the page",
         )
 
@@ -119,7 +119,7 @@ class WebAnalyticsTestCase(CMSTestCase):
 
         self.assertContains(
             response,
-            "googletagmanager",
+            "google-analytics.com",
             msg_prefix="Page should include the Google Analytics snippet code",
         )
         self.assertContains(
@@ -146,16 +146,15 @@ class WebAnalyticsTestCase(CMSTestCase):
         self.assertContains(
             response,
             "dimension4': '" + course_run_link,
-            msg_prefix="Should include course resouce link on the 4th custom dimension",
+            msg_prefix="Should include course resource link on the 4th custom dimension",
         )
         self.assertRegex(
             response_content,
             "dimension5.*" + course_page_title.get("fr"),
             msg="Should include course page title on the 5th custom dimension",
         )
-
         self.assertGreater(
-            response_content.index("googletagmanager"),
+            response_content.index("google-analytics.com"),
             response_content.index("<body"),
             "Web tracking should be at the bottom of the page",
         )
@@ -221,12 +220,12 @@ class WebAnalyticsTestCase(CMSTestCase):
 
         self.assertContains(
             response,
-            "googletagmanager",
+            "google-analytics.com",
             msg_prefix="Page should include the Google Analytics snippet code",
         )
         self.assertGreater(
             response_content.index("<body"),
-            response_content.index("googletagmanager"),
+            response_content.index("google-analytics.com"),
             "Web tracking should be at the head of the page",
         )
         self.assertContains(
@@ -282,4 +281,89 @@ class WebAnalyticsTestCase(CMSTestCase):
             response,
             "dimension3': ''",
             msg_prefix="Should include an empty course run title on the 3rd custom dimension",
+        )
+
+    @override_settings(
+        WEB_ANALYTICS_ID="UA-XXXXXXXXX-X", WEB_ANALYTICS_PROVIDER="google_tag_manager"
+    )
+    def test_web_analytics_google_tag_manager(self):
+        """
+        Test Google Tag Manager has Web Analytics provider for a course page with all the
+        dimensions.
+        """
+        org_page_code = "PUBLIC_ORG"
+        org_page_title = "public title"
+
+        # Create an Organization
+        organization = OrganizationFactory(
+            page_title=org_page_title, should_publish=True, code=org_page_code
+        )
+
+        course_page_code = "XPTO_CODE_FOR_COURSE"
+        course_page_title = {
+            "en": "A fancy title for a course",
+            "fr": "Un titre fantaisiste pour un cours",
+        }
+        course = CourseFactory(
+            page_title=course_page_title,
+            fill_organizations=[organization],
+            code=course_page_code,
+        )
+
+        course_run_link = (
+            "http://example.edx:8073/courses/course-v1:edX+DemoX+01/course/"
+        )
+        course_run_title = "Édition d'automne"
+        course_run_title_safe = "Édition d&#x27;automne"
+        CourseRunFactory(
+            direct_course=course,
+            resource_link=course_run_link,
+            sync_mode="sync_to_public",
+            title=course_run_title,
+            languages=["en", "fr"],
+        )
+
+        course_page = course.extended_object
+        # publish after course run creation so the course run is also published
+        course_page.publish("fr")
+        course_page.publish("en")
+
+        url = course_page.get_absolute_url(language="fr")
+        response = self.client.get(url)
+
+        self.assertContains(
+            response,
+            "googletagmanager.com",
+            msg_prefix="Page should include the Google Tag Manager js code",
+        )
+        self.assertContains(
+            response,
+            "UA-XXXXXXXXX-X",
+            msg_prefix="Page should include the Google Tag Manager tracking id code",
+        )
+        response_content = response.content.decode("UTF-8")
+        self.assertRegex(
+            response_content,
+            "dimension1.*" + org_page_code,
+            msg="Should include organization code on the 1st custom dimension",
+        )
+        self.assertRegex(
+            response_content,
+            "dimension2.*" + course_page_code,
+            msg="Should include course code on the 2nd custom dimension",
+        )
+        self.assertRegex(
+            response_content,
+            "dimension3.*" + course_run_title_safe,
+            msg="Should include course run title on the 3rd custom dimension",
+        )
+        self.assertContains(
+            response,
+            "dimension4': '" + course_run_link,
+            msg_prefix="Should include course resource link on the 4th custom dimension",
+        )
+        self.assertRegex(
+            response_content,
+            "dimension5.*" + course_page_title.get("fr"),
+            msg="Should include course page title on the 5th custom dimension",
         )